+++ src/components/controls/LayerControl.js
@@ -85,7 +85,15 @@ const LayerControl = props => {
 
   const handleEditShape = shape => {
     props.setFocusedShape(shape)
-    props.setActiveDialog('editShape')
+    //props.setActiveDialog('editShape')
+
+    props.setState({
+      ...props.state,
+      dialog: {
+        anchor: null,
+        name: 'editShape',
+      },
+    })
   }
 
   // Only for Circle || Ellipse || Bullseye
@@ -253,9 +261,12 @@ const LayerControl = props => {
               key={`${bullseye.layer}-${bullseye.title}-${bullseye.id}`}
               markerSize={props.markerSize}
               mapZoom={props.mapZoom}
-              setActiveDialog={dialog => props.setActiveDialog(dialog)}
+              //setActiveDialog={dialog => props.setActiveDialog(dialog)}
               setFocusedShape={shape => props.setFocusedShape(shape)}
               tooltipsActive={props.tooltipsActive}
+
+              setState={props.setState}
+              state={props.state}
             />
           ))}
         </LayerGroup>
@@ -272,11 +283,14 @@ const LayerControl = props => {
               markerSize={props.markerSize}
               marker={marker}
               mapZoom={props.mapZoom}
-              setActiveDialog={dialog => props.setActiveDialog(dialog)}
+              //setActiveDialog={dialog => props.setActiveDialog(dialog)}
               setFocusedMarker={marker => props.setFocusedMarker(marker)}
               setFocusedShape={shape => props.setFocusedShape(shape)}
               //toggleEditMarkerDialog={() => props.toggleEditMarkerDialog()}
               tooltipsActive={props.tooltipsActive}
+
+              setState={props.setState}
+              state={props.state}
             />
           ))}
         </LayerGroup>
@@ -293,11 +307,14 @@ const LayerControl = props => {
               markerSize={props.markerSize}
               marker={marker}
               mapZoom={props.mapZoom}
-              setActiveDialog={dialog => props.setActiveDialog(dialog)}
+              //setActiveDialog={dialog => props.setActiveDialog(dialog)}
               setFocusedMarker={marker => props.setFocusedMarker(marker)}
               setFocusedShape={shape => props.setFocusedShape(shape)}
               //toggleEditMarkerDialog={() => props.toggleEditMarkerDialog()}
               tooltipsActive={props.tooltipsActive}
+
+              setState={props.setState}
+              state={props.state}
             />
           ))}
         </LayerGroup>
@@ -314,12 +331,15 @@ const LayerControl = props => {
               markerSize={props.markerSize}
               marker={marker}
               mapZoom={props.mapZoom}
-              setActiveDialog={dialog => props.setActiveDialog(dialog)}
+              //setActiveDialog={dialog => props.setActiveDialog(dialog)}
               setFocusedLatlng={latlng => props.setFocusedLatlng(latlng)}
               setFocusedMarker={marker => props.setFocusedMarker(marker)}
               setFocusedShape={shape => props.setFocusedShape(shape)}
               //toggleEditMarkerDialog={() => props.toggleEditMarkerDialog()}
               tooltipsActive={props.tooltipsActive}
+
+              setState={props.setState}
+              state={props.state}
             />
           ))}
         </LayerGroup>
@@ -336,10 +356,13 @@ const LayerControl = props => {
               markerSize={props.markerSize}
               marker={marker}
               mapZoom={props.mapZoom}
-              setActiveDialog={dialog => props.setActiveDialog(dialog)}
+              //setActiveDialog={dialog => props.setActiveDialog(dialog)}
               setFocusedMarker={marker => props.setFocusedMarker(marker)}
               //toggleEditMarkerDialog={() => props.toggleEditMarkerDialog()}
               tooltipsActive={props.tooltipsActive}
+
+              setState={props.setState}
+              state={props.state}
             />
           ))}
         </LayerGroup>
@@ -356,11 +379,14 @@ const LayerControl = props => {
               markerSize={props.markerSize}
               marker={marker}
               mapZoom={props.mapZoom}
-              setActiveDialog={dialog => props.setActiveDialog(dialog)}
+              //setActiveDialog={dialog => props.setActiveDialog(dialog)}
               setFocusedMarker={marker => props.setFocusedMarker(marker)}
               setFocusedShape={shape => props.setFocusedShape(shape)}
               //toggleEditMarkerDialog={() => props.toggleEditMarkerDialog()}
               tooltipsActive={props.tooltipsActive}
+
+              setState={props.setState}
+              state={props.state}
             />
           ))}
         </LayerGroup>
@@ -378,11 +404,14 @@ const LayerControl = props => {
               markerSize={props.markerSize}
               marker={marker}
               mapZoom={props.mapZoom}
-              setActiveDialog={dialog => props.setActiveDialog(dialog)}
+              //setActiveDialog={dialog => props.setActiveDialog(dialog)}
               setFocusedMarker={marker => props.setFocusedMarker(marker)}
               setFocusedShape={shape => props.setFocusedShape(shape)}
               //toggleEditMarkerDialog={() => props.toggleEditMarkerDialog()}
               tooltipsActive={props.tooltipsActive}
+
+              setState={props.setState}
+              state={props.state}
             />
           ))}
         </LayerGroup>
@@ -399,11 +428,14 @@ const LayerControl = props => {
               markerSize={props.markerSize}
               marker={marker}
               mapZoom={props.mapZoom}
-              setActiveDialog={dialog => props.setActiveDialog(dialog)}
+              //setActiveDialog={dialog => props.setActiveDialog(dialog)}
               setFocusedMarker={marker => props.setFocusedMarker(marker)}
               setFocusedShape={shape => props.setFocusedShape(shape)}
               //toggleEditMarkerDialog={() => props.toggleEditMarkerDialog()}
               tooltipsActive={props.tooltipsActive}
+
+              setState={props.setState}
+              state={props.state}
             />
           ))}
         </LayerGroup>
@@ -420,11 +452,14 @@ const LayerControl = props => {
               markerSize={props.markerSize}
               marker={marker}
               mapZoom={props.mapZoom}
-              setActiveDialog={dialog => props.setActiveDialog(dialog)}
+              //setActiveDialog={dialog => props.setActiveDialog(dialog)}
               setFocusedMarker={marker => props.setFocusedMarker(marker)}
               setFocusedShape={shape => props.setFocusedShape(shape)}
               //toggleEditMarkerDialog={() => props.toggleEditMarkerDialog()}
               tooltipsActive={props.tooltipsActive}
+
+              setState={props.setState}
+              state={props.state}
             />
           ))}
         </LayerGroup>