+++ src/components/dialogs/EditMarkerDrawer.js
@@ -446,7 +446,7 @@ const EditMarkerDrawer = (props) => {
         container={container}
         variant='temporary'
         anchor='left'
-        open={props.open}
+        open={props.state.dialog.name === 'editMarker'}//props.open}
         onClose={props.onClose}
         classes={{ paper: classes.drawerPaper, }}
         ModalProps={{ keepMounted: true, }}