+++ src/components/dialogs/EditShapeDrawer.js
@@ -337,7 +337,7 @@ const ShapeDrawer = (props) => {
         container={container}
         variant='temporary'
         anchor='left'
-        open={props.open}
+        open={props.state.dialog.name === 'editShape'}//props.open}
         onClose={props.onClose}
         classes={{ paper: classes.drawerPaper, }}
         ModalProps={{ keepMounted: true, }}