+++ src/components/layers/Bullseye.js
@@ -84,7 +84,14 @@ const Bullseye = props => {
 
   const handleEditShape = shape => {
     props.setFocusedShape(shape)
-    props.setActiveDialog('editShape')
+    //props.setActiveDialog('editShape')
+    props.setState({
+      ...props.state,
+      dialog: {
+        anchor: null,
+        name: 'editShape',
+      },
+    })
   }
 
   /**